# docs/swagger.yaml
basePath: /api/
definitions:
  chat.Text_message:
    properties:
      message:
        type: string
      tag:
        type: string
    type: object
  chat.response_message:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /:
    post:
      description: Endpoint for testing by returning Hello World
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chat.response_message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chat.response_message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/chat.response_message'
      summary: Testing Server
      tags:
      - Test
  /chat/group:
    post:
      description: Endpoint for setting group to store messages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chat.response_message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chat.response_message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/chat.response_message'
      security:
      - JWT: []
      summary: Set group for messages
      tags:
      - Group
  /chat/message:
    post:
      consumes:
      - application/json
      description: Endpoint for sending messages according to group
      parameters:
      - description: Text tag
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/chat.Text_message'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chat.Text_message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chat.response_message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/chat.response_message'
      security:
      - JWT: []
      summary: Add message according to group
      tags:
      - Message
  /chat/user:
    post:
      description: Endpoint for setting user data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/chat.response_message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/chat.response_message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/chat.response_message'
      security:
      - JWT: []
      summary: Set user data for messages
      tags:
      - User
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"